Remove current-items accounting
It's slow (even if we modify it to use a hash table) and cumbersome and
there is no reason why the user has to always declare existing Items in
every "transaction" where they want to reference them.

Some kind of current-items accounting will reappear later in the RSF
reader because there we have a requirement that RSF files are self
contained. Therefore, any Item referenced by an Entry must be declared
(and currently also defined) in the stream prior to the first reference.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>