Remove digest item-refs
Digest item-refs are only used during RSF parsing and complicate the core
considerably. This patch removes them entirely.

There may be a need in the future to be able to cache the digests for
items. Appropriate code for that will be designed and implemented when
the time comes.

There is now only one type of item-ref: opaque.

 + Reuse the `make-item-ref` symbol for the existing `make-item-ref-opaque`
   procedure.

 + Rename `item-ref-opaque-evict` to `item-ref-evict`.

 + Tighten up argument checking for `make-item`.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>