Rework and rename register-add-item
We rename `register-add-item` to `register-add-item!` because it now
mutates the Items passed to it.

Whilst this procedure is no longer part of the external API, we still
use it internally to ensure Items are actually present in the Backing
Store.

A subsequent patch will update `register-append-entry`.

This procedure calls item-store-add! and immediatly puts the returned
item-ref into the item. This is dangerous because we have no way to take
it out again if the Backing Store transaction subsequently rolls back.
Rolling back the Backing Store transaction  will lead to the following
consequences:

 + Any Items that have been passed to register-add-item! will end up
   with invalid item-refs that could point to an incorrect item or no
   item at all. The item-ref is incredibly unlikely to ever end up
   pointing to the correct item.

 + No further attempts will be made to add these Item objects to the
   Backing Store.

 + If an object, such as an Entry, that depends on these Items is
   successfully added to the Backing Store then that object will be
   corrupt.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>