Rework register-append-entry to use register-add-item!
Call `register-add-item!` on any Items that do not yet have item-refs.

This concludes the patch series that rework the way that Entries and
their Items are written to a Register. Subsequent patches will rework
the RSF support so that it regains the ability to ensure RSF files
always declare or define all the Items that they reference.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>