CLI: `dump` to output a Register as RSF
`orc dump REGISTER`

The `dump` command prints out the entire REGISTER to
STDOUT in RSF format.

This is to allow Registers to be transported around.
In the future, we might also want a mode or command
that takes this output and pushes it to an endpoint
or file, but we can do that manually for now.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>