CLI: `ensure-*` to add entries if key or item is missing
`orc ensure-key  <REGISTER> <REGION> <KEY> [<ITEM-BLOB> ...]`
`orc ensure-item <REGISTER> <REGION> <KEY> [<ITEM-BLOB> ...]`

These commands are similar to `add-entry` in that they may
cause a new entry to be written to the entry log of the
REGISTER with passed REGION and KEY, but they take into account
the existing state of the current records before writing.

The `ensure-key` command first checks the current records
for the KEY. If it is found, there is no further action.
If it is not found, a entry is written as per `add-entry`.
Note that this examines the current record set and tombstones
are considered to not exist - hence if a key has previously
existed but was deleted, a new entry will be written.

The `ensure-item` command first checks the current records
for the KEY. If it is found, it then checks the entry items.
If any of the passed ITEM-BLOBs are not present in the existing
entry, they are added to the existing items and a new entry
is written as per `add-entry`. If a record is found with KEY
and all ITEM-BLOBs are already in it, there is no action.

For both commands, they will print any entries they add
using the current formatter to STDOUT and will remain silent
otherwise.

This is useful when client scripts are converting
an existing dataset into a Register and they don't
want to keep track of what items are in the Register
already; it's easier to just output items again.

E.g. Converting dogs with two colours to Register:
    $ orc colours ensure-key black "{'colour': 'black'}"
    $ orc colours ensure-key brown "{'colour': 'brown'}"
    $ orc dogs ensure-key sausage "{'dog': 'sausage', 'colours': 'black;brown'}"
    $ orc colours ensure-key brown "{'colour': 'brown'}"
    $ orc dogs ensure-key doberman "{'dog': 'doberman', 'colours': 'brown'}"