Add TSV format
This format will output entries and record diffs with each field
separated by tabs. The order of field output is:

    1: region
    2: key
    3: time
    4: item-blob

A line will be printed for each item that the entry contains,
with the entry metadata duplicated. So if an entry contains 2
items, there will be two lines outputted with the same region,
key and time, and differing item blobs.

When outputting diffs, the first field will instead be a + or -
character showing the entry that was previously a record with a
- and the entry that is now a record with a +.

This format is easily parsable with standard GNU tools, such as
`cut`.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>