CLI: `clone` to read and store RSF
`orc clone <LOCATION> <LABEL>`

The `clone` command allows reading a stream of RSF
to create a new Register in a backing store. The
Register is given the internal name specified as
the LABEL parameter.

The LOCATION that RSF is read from can either be
a file or "-" to read from STDIN.

In future versions, the name will come from the
Register's `name` key by default, but we don't
know how to read JSON yet so this is currently
unsupported.

Specifying the LABEL of an existing Register is
an error and no action will be taken.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>