Specify format with the -F option
The `-F` or `--format` option accepts the name
of an entry output format. Recognised entry output
formats can be printed with the `--help` option.

This is in preparation for a future commit which
will introduce further formats for output. The
default format is the TSV format with one line
per item.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>