CLI: `digest` prints Register root hash.
`orc digest REGISTER`

The `digest` command outputs the digest of the REGISTER.

The digest is a string of hexadecimal characters that represents
the current contents of the REGISTER completely. The digest takes
account of all Register history in all regions. If two digests are
the same, then the two Registers they represent are identical.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>