CLI: `add-entry` creates new entry with passed items.
`orc add-entry REGISTER REGION KEY [<ITEM-BLOB> ...]`

The `add-entry` command writes a new entry to the entry log
of a REGISTER with passed REGION, KEY and ITEM-BLOBs. The
timestamp on the entry is the current system time.

Zero or more ITEM-BLOBs can be passed to add multiple items
to the entry. Passing zero items results in a tombstone entry
which effectively deletes the key from the current records.

No processing of any existing entry for the key is done -
if the key exists already, the new entry overrides the old
one in the current records.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>