Add one-line TSV format
This format continues to use tabs to separate fields,
but differs from the existing format in that it will
output all items for an entry on one line as opposed to
one line for each item. This is more appropriate for
use cases which want to process complete entries.

So the field order is:

    1: region
    2: key
    3: time
    4+: item-blobs

If entries contain differing amounts of items, they
will have a different number of fields - entries with
less fields are not padded out with empty fields.

This format can be used by passing "oneline" to the
`-F` option.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>