CLI: `ls` with a key prints entries with that key
`orc ls <REGISTER> <REGION> <KEY>`

The `ls` command prints out all the records in the
REGION of a REGISTER that have a given KEY. Note that
this is printing *records*, so only the current entry
for the key is shown. Keys which have tombstone records
(i.e. entries with no items, signalling deletion) are
not printed using this interface.

REGIONs of a REGISTER are currently either `user` to
print data or `system` to print metadata.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>