CLI: `log` and `diff` commands print history between versions.
`orc log <REGISTER> <START-VERSION> <END-VERSION>`
`orc diff <REGISTER> <START-VERSION> <END-VERSION>`

These commands show the history of a REGISTER between
START-VERSION and END-VERSION.

The `log` command will print all of the entries that
were added to the entry log between the two versions.
This is an operation on entries and not records, so the
even if multiple entries modify the same key, all are shown.

The `diff` command will print the difference in records
between the two versions. Each entry will be prepended with
either a "+" to indicate that it was the new version of a
record for the given key or a "-" to indicate that it was the
old version of a record for a given key. In contrast with `log`
if multiple entries in the history modify the same key
only the version at the start and the version at the end are
shown and the intermediate versions are not printed.

As example output that demonstrates the difference:

    $ orc log my-register ^3 ^0
    user	HELLO	2019-09-10T11:50:24Z	"Hello, World!"
    user	HELLO   2019-09-10T11:50:24Z	"Hello, Simon!"
    user	HELLO	2019-09-10T11:50:24Z	"Goodbye, Simon!"
    $ orc diff my-register ^3 ^0
    -	user	HELLO	2019-09-10T11:50:24Z	"Hello, World!"
    +	user	HELLO	2019-09-10T11:50:24Z	"Goodbye, Simon!"

Version numbers count starting at 0. Caret characters can
be prepended to the version number to count backwards from
the end. So ^0 is the latest version, ^1 is the one before
the latest, and so on. Thus `orc diff register 0 ^0` will
output current records, and `orc log register 0 ^0` will
output full history.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>